backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Data Models & Validation
pydantic==2.9.0
pydantic-settings==2.5.0

# Database
sqlalchemy==2.0.35
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.13.3

# LLM Providers
openai==1.54.0
google-generativeai==0.8.3
google-cloud-aiplatform==1.73.0

# Utilities
python-dotenv==1.0.1
httpx==0.27.2
structlog==24.4.0
python-json-logger==2.0.7
jsonpatch==1.33

# Hashing & Crypto
cryptography==43.0.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0

# Development
ipython==8.29.0
black==24.10.0
ruff==0.7.3
